Moved testing out to unittest folder
