+++ scripts/testing/unittests/grabcut-result-apple.png
